Add no_std support, Unit enum, min_unit feature, and move tests to tests/
