+++ tests/common/mod.rs
@@ -1,0 +1,11 @@
+#![allow(dead_code)]
+
+pub const MIN: u64 = 60;
+pub const HOUR: u64 = 60 * MIN;
+pub const DAY: u64 = 24 * HOUR;
+pub const WEEK: u64 = 7 * DAY;
+pub const MONTH: u64 = 2_629_846;
+pub const YEAR: u64 = 31_558_150;
+pub const KILO_YEAR: u64 = 1_000 * YEAR;
+pub const MEGA_YEAR: u64 = 1_000 * KILO_YEAR;
+pub const GIGA_YEAR: u64 = 1_000 * MEGA_YEAR;
